Fixed warnings
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@80500 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/blastdb/makeblastdb.cpp
@@ -1119,7 +1119,7 @@ void CMakeBlastDBApp::x_BuildDatabase()
     // Max file size
 
     Uint8 bytes = NStr::StringToUInt8_DataSize(args["max_file_sz"].AsString());    
-    if (bytes >= (1L << 32)) {
+    if (bytes >= (Uint8)(1L << 32)) {
         NCBI_THROW(CInvalidDataException, eInvalidInput,
                 "max_file_sz must be < 4 GiB");
     }
